requirements.txt
neo4j
langchain
langchain-openai
langchain-core
python-dotenv
